% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{abcoxph_setup}
\alias{abcoxph_setup}
\title{Set up the Cox Proportional Hazard Model that you want to infer.}
\usage{
abcoxph_setup(formula, cens, data, prior_control = NULL, RW2BINS = NULL)
}
\arguments{
\item{formula}{A formula specifies your observed times as response variable, regressors type are smoothing, id, or linear fixed effect.}

\item{cens}{A STRING specifies the name of your censoring indicator in the dataframe.}

\item{data}{Target dataframe.}

\item{prior_control}{A list specifies your prior distributions.}

\item{RW2BINS}{A number specifies the number of bins you want to use for smoothing.}
}
\value{
A list that contains all the information needed for your model to be fitted using abcoxph_fit.
}
\description{
Setting up a target CoxPH Model, by specifying the formula. The response in the variable should be your
observed times, and the regressors's type are well identified with s denote smoothing and id denote between-subject
index, other terms are assumed to be linear fixed effect.
}
